requirements.txt
# ── Web framework ────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9      # required for FastAPI file uploads

# ── Frontend ──────────────────────────────────────────────────────────────────
streamlit==1.35.0
Pillow==10.3.0

# ── YOLO detection ───────────────────────────────────────────────────────────
ultralytics==8.2.18          # YOLOv8; pulls torch CPU automatically if no GPU

# ── RAG stack ────────────────────────────────────────────────────────────────
sentence-transformers==3.0.1
faiss-cpu==1.8.0

# ── HTTP client (for Ollama) ──────────────────────────────────────────────────
requests==2.32.3

# ── Utilities ────────────────────────────────────────────────────────────────
numpy==1.26.4
ollama

